refactor(cli): extract task detail view
